// src/@types/models/status.dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

interface stats {
  count: number;
  percentage: number;
}
/**
 *
 *
 * @export
 * @interface StatsDto
 */
export interface StatsDto {
  user: stats;

  vendor: stats;

  supplier: stats;
}